% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concentration.observed}
\alias{concentration.observed}
\title{Market concentration observed by region and year}
\format{
An object of class \code{data.frame} with 5 rows and 10 columns.
}
\usage{
concentration.observed
}
\description{
A normalized Herfindahl-Hirschman index of concentration (Hirschman 1964) by 
region for DAP/MAP fertilizer trade volumes 2013-2017.
}
\details{
The index of concentration takes the value 0 when a regional market is 
provided equally by all six suppliers (international suppliers + 
domestic industry), and approaches 1 when the market in question has 
a single supplier.

The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

Regional classification includes: Africa, East Asia, Eastern Europe and Central Asia, Latin America, North America, Oceania, South Asia, Western and Central Europe, and Western Asia.
}
\references{
\itemize{
  \item Hirschman, A. (1964). The Paternity of an Index. The American Economic Review, 54(5): 761.
}
}
\keyword{datasets}
