% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concentration.simulated}
\alias{concentration.simulated}
\title{Market concentration simulated by region and scenario}
\format{
An object of class \code{data.frame} with 2000 rows and 11 columns.
}
\usage{
concentration.simulated
}
\description{
A normalized Herfindahl-Hirschman index of concentration (Hirschman 1964) by
region for DAP/MAP fertilizer trade volumes in 2030 scenarios.
}
\details{
The index of concentration takes the value 0 when a regional market is 
provided equally by all six suppliers (international suppliers + 
domestic industry), and approaches 1 when the market in question has 
a single supplier.

The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

The compilation covers 1000 bootstrap simulations by distributed DAP/MAP 
market model by FAO "Business as Usual" and "Stratified Societies" 
scenarios year 2030.

Regional classification includes: Africa, East Asia, Eastern Europe and Central Asia, Latin America, North America, Oceania, South Asia, Western and Central Europe, and Western Asia.
}
\references{
\itemize{
  \item Hirschman, A. (1964). The Paternity of an Index. The American Economic Review, 54(5): 761.
  \item FAO. 2018. The future of food and agriculture - Alternative pathways to 2050. Rome.
}
}
\keyword{datasets}
