% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{dapmap.model}
\alias{dapmap.model}
\title{Measures of competition for distributed DAP/MAP market model}
\seealso{
\code{\link{concentration.observed}}, 
\code{\link{diversification.observed}},
\code{\link{self_sufficiency.observed}}, 
\code{\link{concentration.simulated}}, 
\code{\link{diversification.simulated}},
\code{\link{self_sufficiency.simulated}}
}
