% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diversification.simulated}
\alias{diversification.simulated}
\title{Supplier diversification observed by region and scenario}
\format{
An object of class \code{data.frame} with 2000 rows and 7 columns.
}
\usage{
diversification.simulated
}
\description{
An inverse normalized Herfindahl-Hirschman index of diversification 
(Berry 1971) by international supplier for DAP/MAP fertilizer trade volumes 
in 2030 scenarios.
}
\details{
The index of diversification takes the value 0 when an international supplier
is active on a single market, and approaches 1 when the supplier in question 
sells in equal shares to all markets.

The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

The compilation covers 1000 bootstrap simulations by distributed DAP/MAP 
market model by FAO "Business as Usual" and "Stratified Societies" 
scenarios year 2030.

This compilation covers the following phosphate companies and groups of companies:
\itemize{
  \item Chinese exporters
  \item Maaden https://www.maaden.com.sa/
  \item Mosaic http://www.mosaicco.com/
  \item PhosAgro/EuroChem https://phosagro.com/, https://www.eurochemgroup.com/
  \item OCP https://www.ocpgroup.ma/
}
}
\references{
\itemize{
  \item Berry, C. (1971). Corporate Growth and Diversification. Journal of Law and Economics, 14(2): 371-383.
  \item FAO. 2018. The future of food and agriculture - Alternative pathways to 2050. Rome.
}
}
\keyword{datasets}
