% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_sufficiency.observed}
\alias{self_sufficiency.observed}
\title{The market share of domestic supply observed by region and year}
\format{
An object of class \code{data.frame} with 5 rows and 10 columns.
}
\usage{
self_sufficiency.observed
}
\description{
A market share of domestic supply by region for DAP/MAP fertilizer trade volumes 2013-2017.
}
\details{
The market share of local suppliers takes the value 0 when a regional market is 
provided only by international suppliers, and approaches 1 when the market in question has 
only domestic supply.

The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

Regional classification includes: Africa, East Asia, Eastern Europe and Central Asia, Latin America, North America, Oceania, South Asia, Western and Central Europe, and Western Asia.
}
\keyword{datasets}
