% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{self_sufficiency.simulated}
\alias{self_sufficiency.simulated}
\title{The market share of domestic supply observed by region and scenario}
\format{
An object of class \code{data.frame} with 2000 rows and 11 columns.
}
\usage{
self_sufficiency.simulated
}
\description{
A market share of domestic supply by region for DAP/MAP fertilizer trade 
volumes in 2030 scenarios.
}
\details{
The market share of local suppliers takes the value 0 when a regional market is 
provided only by international suppliers, and approaches 1 when the market in question has 
only domestic supply.

The following definitions apply: DAP - diammonium phosphate, MAP - monoammonium phosphate.

The compilation covers 1000 bootstrap simulations by distributed DAP/MAP 
market model by FAO "Business as Usual" and "Stratified Societies" 
scenarios year 2030.

Regional classification includes: Africa, East Asia, Eastern Europe and Central Asia, Latin America, North America, Oceania, South Asia, Western and Central Europe, and Western Asia.
}
\keyword{datasets}
